src/rest/requirements.txt
Django==1.8.6
django-tastypie==0.12.2
celery==3.1.19
django-celery==3.1.17
django-redis==4.3.0
redis==2.10.5
Pillow==2.3.0
